No.22とNo.23実装完了

+++ src/main/java/com/example/entity/ProductWithCategoryName.java
@@ -19,11 +19,12 @@ public class ProductWithCategoryName {
 
 	private Integer height;
 
-	private Integer price;
+	private double price;
 
 	private String categoryName;
 
-	public ProductWithCategoryName(Long id, String code, String name, Integer weight, Integer height, Integer price, String categoryName) {
+	public ProductWithCategoryName(Long id, String code, String name, Integer weight, Integer height, double price,
+			String categoryName) {
 		this.setId(id);
 		this.setCode(code);
 		this.setName(name);
